Implement TwoPhaseFileHeader details output in pg_xlogdump
Output the following in pg_xlogdump for prepare transaction records:
* Global transaction ID
* Prepare time
* Tablespace OIDs:
	* 'tablespace_oid_to_delete_on_abort'
	* 'tablespace_oid_to_delete_on_commit'

These data come from TwoPhaseFileHeader structure which is written
in xlog in GPDB.

As TwoPhaseFileHeader structure is a "private" one defined in
twophase.c, a new function to print its contents is defined in that
file and declared in twophase.h.

As pg_xlogdump is a standalone application, preprocessor directives
are introduced into twophase.c in order to preserve this build
independence.

Co-Authored-By: David Kimura <dkimura@pivotal.io>